+++ src/types/ISignin.ts
@@ -10,7 +10,7 @@ export interface IServerUser {
   profileImgUrl: string;
   accessToken: string;
   refreshToken: string;
-  role: 'ROLE_USER' | 'ROLE_SELLER';
+  role: 'ROLE_USER' | 'ROLE_SELLER' | 'ROLE_ADMIN';
   platform: 'BASIC' | 'KAKAO';
 }
